chore(native): Compile jinja in fallback build
chore: fix

chore: fix

chore: fix

chore: another way

chore: fmt

chore: fmt
