+++ packages/cubejs-backend-native/src/python/mod.rs
@@ -1,10 +1,8 @@
-pub(crate) mod cross;
 pub(crate) mod cube_config;
 mod entry;
 #[cfg(target_os = "linux")]
 pub(crate) mod linux_dylib;
 pub(crate) mod python_model;
 pub(crate) mod runtime;
-pub mod template;
 
 pub use entry::python_register_module;
